+++ wasatch/Reading.py
@@ -41,6 +41,7 @@ class Reading:
         self.elapsed_from_request      = None
         self.elapsed_since_last        = None
         self.keep_alive                = False
+        self.image_format              = None
 
         # currently only populated by AutoRaman
         self.new_integration_time_ms   = None